# docs/api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A multiplayer FPS game server with OAuth authentication and N:1 account linking",
        "title": "LIFE Game Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/homveloper/life",
            "email": "support@lifegame.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/homveloper/life/blob/main/LICENSE"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/api/v1/trainer.Create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new trainer with nickname for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "Create a new trainer",
                "parameters": [
                    {
                        "description": "JSON-RPC request with CreateTrainerRequest params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_CreateTrainerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created trainer information",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_CreateTrainerResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/trainer.Get": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get trainer information for the authenticated user (auto-creates if not exists)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "Get trainer information",
                "parameters": [
                    {
                        "description": "JSON-RPC request with GetTrainerRequest params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_GetTrainerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainer information",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_GetTrainerResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/trainer.List": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all trainers in the game world (currently returns mock data)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "List all trainers",
                "parameters": [
                    {
                        "description": "JSON-RPC request with ListTrainerRequest params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_ListTrainerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of trainers",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_ListTrainerResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/trainer.Move": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Move trainer to a new position on the game map with coordinate validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "Move trainer to new position",
                "parameters": [
                    {
                        "description": "JSON-RPC request with MoveTrainerRequest params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_MoveTrainerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated trainer with new position",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_MoveTrainerResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters or invalid coordinates",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/trainer.Status": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed status information for the authenticated trainer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainer"
                ],
                "summary": "Get trainer status",
                "parameters": [
                    {
                        "description": "JSON-RPC request with GetTrainerRequest params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_GetTrainerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainer status information",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_StatusTrainerResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_danghamo_life_internal_api_jsonrpcx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_CreateTrainerRequest": {
            "type": "object",
            "properties": {
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string"
                },
                "params": {
                    "$ref": "#/definitions/internal_api_handlers.CreateTrainerRequest"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_GetTrainerRequest": {
            "type": "object",
            "properties": {
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string"
                },
                "params": {
                    "$ref": "#/definitions/internal_api_handlers.GetTrainerRequest"
                }
            }
        },
...
        "github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_ListTrainerRequest": {
            "type": "object",
            "properties": {
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string"
                },
                "params": {
                    "$ref": "#/definitions/internal_api_handlers.ListTrainerRequest"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.RequestT-internal_api_handlers_MoveTrainerRequest": {
            "type": "object",
            "properties": {
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string"
                },
                "params": {
                    "$ref": "#/definitions/internal_api_handlers.MoveTrainerRequest"
                }
            }
        },
...
        "github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_CreateTrainerResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "$ref": "#/definitions/internal_api_handlers.CreateTrainerResponse"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_GetTrainerResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "$ref": "#/definitions/internal_api_handlers.GetTrainerResponse"
                }
            }
        },
...
        "github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_ListTrainerResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "$ref": "#/definitions/internal_api_handlers.ListTrainerResponse"
                }
            }
        },
        "github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_MoveTrainerResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "$ref": "#/definitions/internal_api_handlers.MoveTrainerResponse"
                }
            }
        },
...
        "github_com_danghamo_life_internal_api_jsonrpcx.ResponseT-internal_api_handlers_StatusTrainerResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_danghamo_life_internal_api_jsonrpcx.JSONRPCError"
                },
                "id": {},
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "$ref": "#/definitions/internal_api_handlers.StatusTrainerResponse"
                }
            }
        },
...
        "internal_api_handlers.CreateTrainerRequest": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.GetTrainerRequest": {
            "type": "object"
        },
...
        "internal_api_handlers.ListTrainerRequest": {
            "type": "object"
        },
        "internal_api_handlers.ListTrainerResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "trainers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_handlers.TrainerSummary"
                    }
                }
            }
        },
        "internal_api_handlers.MoveTrainerRequest": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api_handlers.TrainerSummary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "position": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}